+++ validator/src/main/java/org/mustangproject/validator/PDFValidator.java
@@ -220,7 +220,7 @@ public class PDFValidator extends Validator {
 			}
 			boolean documentFilenameValid=false;
 			for (int i = 0; i < nodes.getLength(); i++) {
-				String[] valueArray = { "factur-x.xml", "ZUGFeRD-invoice.xml", "zugferd-invoice.xml" };
+				String[] valueArray = { "factur-x.xml", "ZUGFeRD-invoice.xml", "zugferd-invoice.xml", "xrechnung.xml" };
 				if (stringArrayContains(valueArray, nodes.item(i).getTextContent())) {
 					documentFilenameValid=true;
 				}